pyproject.toml
[project]
name = "event-booking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "pydantic-settings>=2.9.1",
    "django-stubs[compatible-mypy]>=5.2.0",
    "djangorestframework-simplejwt>=5.5.0",
    "django-silk>=5.3.2",
    "redis>=6.2.0",
    "django-redis>=5.4.0",
]
